[REFACTOR] : ♻️ Server에서 받은 데이터 가공 hooks 생성
